abstracted command executors
